pyproject.toml
[tool.poetry]
name = "tum-proj"
version = "0.1.0"
description = ""
authors = ["MikeTsvirkunov <mike.tsvirkunov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.1"
requests = "^2.31.0"
fastapi = "^0.110.2"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
python-ioc = "^1.13.0"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.8.3"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
